[refactor] Move QuantizedLayerPlanner to layerplanner package root-level
